[plugin-query-and-mutation] resolve sonar cloud errors
