+++ bigbluebutton-html5/imports/ui/components/plugins-engine/data-consumption/error-boundary/handler.tsx
@@ -1,56 +1,11 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import * as React from 'react';
 import { makeCustomHookIdentifierFromArgs } from 'bigbluebutton-html-plugin-sdk/dist/cjs/data-consumption/utils';
-import { UpdatedEventDetails } from 'bigbluebutton-html-plugin-sdk/dist/cjs/core/types';
-import { HookEvents } from 'bigbluebutton-html-plugin-sdk/dist/cjs/core/enum';
 import { DataConsumptionHooks } from 'bigbluebutton-html-plugin-sdk/dist/cjs/data-consumption/enums';
-import { CustomSubscriptionArguments } from 'bigbluebutton-html-plugin-sdk/dist/cjs/data-consumption/domain/shared/custom-subscription/types';
-import { CustomQueryArguments } from 'bigbluebutton-html-plugin-sdk/dist/cjs/data-consumption/domain/shared/custom-query/types';
-import { CustomDataConsumptionHooksErrorBoundaryProps, DataConsumptionFallbackHandlerProps, ErrorInformation } from './types';
+import { CustomDataConsumptionHooksErrorBoundaryProps } from './types';
 import ErrorBoundary from '../../../common/error-boundary/component';
-import { deleteSubscriptionHookEntry } from '../utils';
-
-const fallbackHandler = (props: DataConsumptionFallbackHandlerProps) => {
-  const {
-    hook, errorInformation, setDataConsumptionHookWithArgumentUtilizationCount,
-  } = props;
-  const customSubscriptionData = {
-    data: null,
-    error: errorInformation,
-    loading: false,
-  };
-
-  const updateCustomSubscriptionForPlugin = () => {
-    window.dispatchEvent(
-      new CustomEvent<UpdatedEventDetails<any>>(
-        HookEvents.BBB_CORE_SENT_NEW_DATA,
-        {
-          detail: {
-            data: customSubscriptionData,
-            hook,
-            hookArguments: {
-              query: errorInformation.dataConsumptionInformation.query,
-              variables: errorInformation.dataConsumptionInformation.variables,
-            } as CustomSubscriptionArguments | CustomQueryArguments,
-          } as UpdatedEventDetails<any>,
-        },
-      ),
-    );
-  };
-
-  React.useEffect(() => {
-    const hookArguments = {
-      query: errorInformation.dataConsumptionInformation.query,
-      variables: errorInformation.dataConsumptionInformation.variables,
-    };
-    updateCustomSubscriptionForPlugin();
-    deleteSubscriptionHookEntry(
-      setDataConsumptionHookWithArgumentUtilizationCount,
-      hook, hookArguments,
-    );
-  }, []);
-  return null;
-};
+import FallbackHandler from './fallback-handler/handler';
+import { ErrorInformation } from './fallback-handler/types';
 
 const CustomDataConsumptionHooksErrorBoundary: React.FC<CustomDataConsumptionHooksErrorBoundaryProps> = (
   props: CustomDataConsumptionHooksErrorBoundaryProps,
@@ -94,11 +49,13 @@ const CustomDataConsumptionHooksErrorBoundary: React.FC<CustomDataConsumptionHoo
     <ErrorBoundary
       key={makeCustomHookIdentifierFromArgs(hookWithArguments.hookArguments)}
       Fallback={() => {
-        fallbackHandler({
-          errorInformation,
-          hook: dataConsumptionHook,
-          setDataConsumptionHookWithArgumentUtilizationCount,
-        });
+        return (
+          <FallbackHandler
+            errorInformation={errorInformation}
+            hook={dataConsumptionHook}
+            setDataConsumptionHookWithArgumentUtilizationCount={setDataConsumptionHookWithArgumentUtilizationCount}
+          />
+        );
       }}
       logMetadata={logMetadata}
       errorMessage={errorMessage}
